mrc_api_logstash/requirements.txt
transformers==3.3.1
flask==2.1.2
flask-restful==0.3.9
gunicorn
python3-logstash